requirements.txt
langchain==0.1.16
langchain-community==0.0.33
sentence-transformers==2.2.2
huggingface_hub==0.16.4
qdrant-client==1.9.0
gigachat==0.1.42
python-dotenv==1.0.0